+++ Tests/Core/Detour/DetourTests.cpp
@@ -2,6 +2,8 @@
 #include <Zyntercept/Core/Detour/Detour.h>
 #include <Zyntercept/Core/Syscall/Syscall.h>
 
+#include <cmath>
+
 #if defined(ZYNTERCEPT_WINDOWS)
 #include <Windows.h>
 static ZyanVoidPointer ProcessIdentifier = GetCurrentProcess();
@@ -9,7 +11,7 @@ static ZyanVoidPointer ProcessIdentifier = GetCurrentProcess();
 
 #if defined(ZYNTERCEPT_UNIX)
 #include <unistd.h>
-static ZyanVoidPointer ProcessIdentifier = (ZyanVoidPointer)getpid();
+static ZyanVoidPointer ProcessIdentifier = (ZyanVoidPointer)(uint64_t)getpid();
 #endif
 
 static void NoopFunction() {
@@ -24,7 +26,7 @@ static void NoopFunction() {
 // We know that any integer number can be written in the form of 6k+i, where k is a non-negative integer (like 0, 1, 2, 3,...) 
 // and i is a number between 0 and 5 (so i can be 0, 1, 2, 3, 4, or 5).
 
-// If we look closely, we’ll notice that when i is 0, 2, 3, or 4, the numbers 6k, 6k + 2, 6k + 3, and 6k + 4 are all divisible by
+// If we look closely, weï¿½ll notice that when i is 0, 2, 3, or 4, the numbers 6k, 6k + 2, 6k + 3, and 6k + 4 are all divisible by
 // either 2 or 3. But prime numbers greater than 3 can't be divisible by 2 or 3. Therefore, the only forms left that a prime 
 // number can have are 6k + 1 or 6k + 5 (since these forms are not divisible by 2 or 3).
